split build command host effect fixtures
